map indirectly through <Plug>... mappings
Avoid mapping keys directly with `noremap`-type commands.  Create new mappings starting with `<Plug>MatchIt`... and map key to them to be remapped.

This style is similar to that of *unimpaired* and *extended-ft* plugins.  It allows the operations to be accessible even if the '%' key is remapped by another configuration script.